fix shared libs installer on fedora

+++ roles/shared_libs/templates/install.py
@@ -4,7 +4,18 @@ import site
 import os
 
 src = "{}/smartserver".format(os.path.dirname(__file__))
-dest = "{}/smartserver".format(site.getsitepackages()[0])
 
-if not os.path.exists(dest):
-    os.symlink(src, dest)
+site_packages = site.getsitepackages()
+site_packages.sort()
+
+for dir in site_packages:
+    
+    if not os.path.exists(dir):
+        continue
+    
+    dest = "{}/smartserver".format(dir)
+
+    if not os.path.exists(dest):
+        os.symlink(src, dest)
+        
+    break
